# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Bioskuy API
  description: API for Bioskuy movie ticket booking system
  version: 1.0.0
...
paths:
  /users:
    post:
      summary: Register a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"
              password: "securePassword123"
              phoneNumber: "+1234567890"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
              example:
                message: "User created successfully"
                data:
                  id: 1
                  name: "John Doe"
                  email: "john.doe@example.com"
                  phoneNumber: "+1234567890"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              example:
                message: "Invalid input: Email is already in use"
                data: null
  /users/me:
    get:
      summary: Get current logged in user
      tags:
        - Users
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
              example:
                message: "Current user retrieved successfully"
                data:
                  name: "John Doe"
                  email: "john.doe@example.com"
                  phoneNumber: "+1234567890"
        '404':
          description: Current user not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              example:
                message: "Current user not found: User not found with email: example@email.com"
                data: null
  /users/{id}:
    put:
      summary: Update user profile
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"
              password: "securePassword123"
              phoneNumber: "+1234567890"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
                example:
                  message: "User updated successfully"
                  data:
                    id: 1
                    name: "John Doe"
                    email: "john.doe@example.com"
                    password: "securePassword123"
                    phoneNumber: "+1234567890"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                example:
                  message: "User not found"
                  data: null
  /users/login:
    post:
      summary: User login
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
            example:
              email: "john.doe@example.com"
              password: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
              example:
                message: "Login successful"
                data:
                  id: 1
                  name: "John Doe"
                  email: "john.doe@example.com"
                  phoneNumber: "+1234567890"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              example:
                message: "Invalid credentials: Email or password is incorrect"
                data: null
  /users/logout:
    post:
      summary: User logout
      tags:
        - Users
      parameters:
        - name: Authorization
          in: header
          description: JWT token in the format "Bearer {token}"
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              example:
                message: "Logout successful"
                data: null
        '400':
          description: No authentication token provided
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              example:
                message: "No authentication token provided"
                data: null
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              example:
                message: "An error occurred during logout: {error message}"
                data: null
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
        data:
          nullable: true
          oneOf:
              - type: array
              - type: object
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phoneNumber:
          type: string
      required:
        - name
        - email
        - password
...